fixed logging for gzip and refactored location of helper functions
