+++ modules/helpers.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"regexp"
 	"strings"
 	"time"
 
@@ -158,28 +159,6 @@ func getRequestMethod(repl *caddy.Replacer) (string, error) {
 	return methodStr, nil
 }
 
-// ensureUniqueProviderHost ensures the provider has a unique host in the network's providers map
-// by appending a counter if necessary. Returns the unique host value.
-func (d *DinMiddleware) ensureUniqueProviderHost(networkName string, host string) string {
-	// Get existing hosts with the same base name
-	baseHosts := []string{}
-
-	for existingHost := range d.Networks[networkName].Providers {
-		// We need to match exact host or host-N pattern
-		if existingHost == host || strings.HasPrefix(existingHost, host+"-") {
-			baseHosts = append(baseHosts, existingHost)
-		}
-	}
-
-	// If no hosts with this base exist yet, use base host without suffix
-	if len(baseHosts) == 0 {
-		return host
-	}
-
-	// For subsequent hosts, use host-1, host-2, etc.
-	return fmt.Sprintf("%s-%d", host, len(baseHosts))
-}
-
 // decompressGzipBodyIfNecessary checks if the body is gzipped based on headers
 // and attempts to decompress it. It returns the processed body (decompressed or original).
 func decompressGzipBodyIfNecessary(headers http.Header, bodyBytes []byte, lg *logger.LoggerClient, networkPath string) []byte {
@@ -203,6 +182,183 @@ func decompressGzipBodyIfNecessary(headers http.Header, bodyBytes []byte, lg *lo
 		}
 	}
 	return bodyBytes
+}
+
+// LogFailedAttemptParams holds all the configuration data needed for logging failed attempts
+type LogFailedAttemptParams struct {
+	Reason              string
+	Logger              *logger.LoggerClient
+	NetworkPath         string
+	FailedAttemptNumber int // 1-indexed
+	MaxAttempts         int
+	StatusCodeOfFailure int
+	Error               error                   // Can be nil
+	Replacer            *caddy.Replacer         // Caddy replacer to get context data
+	ParsedReqBody       *dinHttp.JSONRPCRequest // Parsed request body
+	RawResponseBody     []byte                  // Raw response body to parse internally
+}
+
+// Helper function to log a failed request attempt that will be retried.
+// This function is intended to be run as a goroutine.
+func logFailedAttempt(params *LogFailedAttemptParams) {
+	// --- Extract data within the async function ---
+	provider := "unknown"
+	if provVal, provOk := params.Replacer.Get(RequestProviderKey); provOk {
+		if pStr, strOk := provVal.(string); strOk {
+			provider = pStr
+		}
+	}
+
+	var requestMethod string
+	var requestParams json.RawMessage
+	rawRequestBodySnippet := ""
+
+	// Always extract raw request body for logging
+	if v, okGet := params.Replacer.Get(RequestBodyKey); okGet {
+		if bodyBytes, okCast := v.([]byte); okCast && len(bodyBytes) > 0 {
+			// Use a helper for min if not available; assuming min is defined elsewhere or use direct comparison
+			length := 500 // Increased from 100 to 500 for more context
+			if len(bodyBytes) < length {
+				length = len(bodyBytes)
+			}
+			rawRequestBodySnippet = string(bodyBytes[:length])
+		}
+	}
+
+	if params.ParsedReqBody != nil {
+		requestMethod = params.ParsedReqBody.Method
+		if len(params.ParsedReqBody.Params) > 0 {
+			requestParams = params.ParsedReqBody.Params
+		}
+	} else if rawRequestBodySnippet != "" {
+		// Try to extract method from raw request body if parsing failed
+		var tempReq struct {
+			Method string `json:"method"`
+		}
+		if v, okGet := params.Replacer.Get(RequestBodyKey); okGet {
+			if bodyBytes, okCast := v.([]byte); okCast && len(bodyBytes) > 0 {
+				if err := json.Unmarshal(bodyBytes, &tempReq); err == nil && tempReq.Method != "" {
+					requestMethod = tempReq.Method
+				} else {
+					// If JSON parsing fails, try to extract method using regex as fallback
+					methodRegex := regexp.MustCompile(`"method"\s*:\s*"([^"]+)"`)
+					if matches := methodRegex.FindSubmatch(bodyBytes); len(matches) > 1 {
+						requestMethod = string(matches[1])
+					}
+				}
+			}
+		}
+	}
+
+	// Parse response body to extract error information
+	var jsonRPCError *dinHttp.JSONRPCError
+	var rawResponseSnippet string
+	var parsedResponseType string
+
+	if len(params.RawResponseBody) > 0 {
+		// Decompress gzipped response body if necessary
+		processedResponseBody := params.RawResponseBody
+		if len(params.RawResponseBody) >= 2 && params.RawResponseBody[0] == 0x1f && params.RawResponseBody[1] == 0x8b {
+			// Create headers to indicate gzip encoding for the decompression function
+			headers := make(http.Header)
+			headers.Set("Content-Encoding", "gzip")
+			processedResponseBody = decompressGzipBodyIfNecessary(headers, params.RawResponseBody, params.Logger, params.NetworkPath)
+		}
+
+		// Create a snippet of the processed response for logging
+		length := 500
+		if len(processedResponseBody) < length {
+			length = len(processedResponseBody)
+		}
+		rawResponseSnippet = string(processedResponseBody[:length])
+
+		// Try to parse as JSON-RPC response first using the processed (potentially decompressed) body
+		var jsonRPCResponse dinHttp.JSONRPCResponse
+		if err := json.Unmarshal(processedResponseBody, &jsonRPCResponse); err == nil {
+			parsedResponseType = "jsonrpc"
+			if jsonRPCResponse.Error != nil {
+				jsonRPCError = jsonRPCResponse.Error
+			}
+		} else {
+			// Try to parse as generic JSON to see if it's valid JSON
+			var genericJSON interface{}
+			if err := json.Unmarshal(processedResponseBody, &genericJSON); err == nil {
+				parsedResponseType = "json"
+			} else {
+				parsedResponseType = "raw"
+			}
+		}
+	}
+	// --- End data extraction ---
+
+	logFields := []zap.Field{
+		zap.String("network", params.NetworkPath),
+		zap.String("provider", provider),
+		zap.Int("failed_attempt_number", params.FailedAttemptNumber),
+		zap.Int("max_attempts", params.MaxAttempts),
+		zap.Int("status_code", params.StatusCodeOfFailure),
+		zap.String("reason", params.Reason),
+	}
+
+	if params.Error != nil {
+		logFields = append(logFields, zap.NamedError("error", params.Error))
+	}
+
+	// Add JSON-RPC error information if present
+	if jsonRPCError != nil {
+		logFields = append(logFields,
+			zap.Int("jsonrpc_error_code", jsonRPCError.Code),
+			zap.String("jsonrpc_error_message", jsonRPCError.Message))
+		if jsonRPCError.Data != nil {
+			logFields = append(logFields, zap.Any("jsonrpc_error_data", jsonRPCError.Data))
+		}
+	}
+
+	if requestMethod != "" {
+		logFields = append(logFields, zap.String("request_method", requestMethod))
+	}
+
+	if len(requestParams) > 0 && string(requestParams) != "null" {
+		var decodedParams interface{}
+		if errUnmarshal := json.Unmarshal(requestParams, &decodedParams); errUnmarshal == nil {
+			logFields = append(logFields, zap.Any("request_params", decodedParams))
+		} else {
+			logFields = append(logFields, zap.String("raw_request_params", string(requestParams)))
+			params.Logger.Debug("Async retry error log: Failed to unmarshal requestParams for structured logging, logging as raw string.",
+				zap.Error(errUnmarshal),
+				zap.String("raw_params_attempted", string(requestParams)))
+		}
+	}
+
+	// Always log raw request body snippet if available
+	if rawRequestBodySnippet != "" {
+		// Try to parse as JSON for structured logging
+		var requestBodyJSON interface{}
+		if err := json.Unmarshal([]byte(rawRequestBodySnippet), &requestBodyJSON); err == nil {
+			logFields = append(logFields, zap.Any("raw_request_body", requestBodyJSON))
+		} else {
+			// If not valid JSON, log as string
+			logFields = append(logFields, zap.String("raw_request_body", rawRequestBodySnippet))
+		}
+	}
+
+	// Log response information
+	if rawResponseSnippet != "" {
+		// Try to parse as JSON for structured logging
+		var responseBodyJSON interface{}
+		if err := json.Unmarshal([]byte(rawResponseSnippet), &responseBodyJSON); err == nil {
+			logFields = append(logFields,
+				zap.Any("raw_response_body", responseBodyJSON),
+				zap.String("response_type", parsedResponseType))
+		} else {
+			// If not valid JSON, log as string
+			logFields = append(logFields,
+				zap.String("raw_response_body", rawResponseSnippet),
+				zap.String("response_type", parsedResponseType))
+		}
+	}
+
+	params.Logger.Warn("Request attempt failed, initiating retry", logFields...)
 }
 
 // PostRequestTaskParams holds all parameters for handlePostRequestTasks.
@@ -327,88 +483,6 @@ func handlePostRequestTasks(params PostRequestTaskParams) {
 		HealthStatus:   healthStatus,
 		Environment:    string(params.DinMiddleware.Env),
 	}, params.Duration, params.ParsedReqBody)
-}
-
-// processHCMethodResponseAsync asynchronously processes responses for health check method requests.
-// It extracts the request method from the replacer, verifies if it matches the network's health check
-// method (HCMethod), and if so, processes the block number from the response body.
-// This function is designed to run in a separate goroutine to avoid blocking the main request handling flow.
-// When a valid block number is extracted, it also retrieves the corresponding block hash and adds both
-// to the network's block history. This information is crucial for tracking the network's current state
-// and ensuring proper synchronization across providers.
-func (d *DinMiddleware) processHCMethodResponseAsync(networkObj *network, networkPath string, respBody []byte, respStatus int, method string) {
-	if len(respBody) == 0 || networkObj == nil || networkObj.HCMethod == "" {
-		return
-	}
-
-	// If method is "other_method" and networkObj.HCMethod is "eth_blockNumber", this should be true.
-	if method != networkObj.HCMethod {
-		return
-	}
-
-	// This log should only appear if Condition 2 is false.
-	d.logger.Debug("Goroutine: Processing response for HCMethod", zap.String("method", method), zap.String("network", networkPath))
-
-	// Create synthetic request context for consistent logging
-	// Since this is async processing of a response, we use "din" as provider identifier
-	providerHost := "din"
-	repl, jsonRPCReq, payload := createHealthCheckRequestContext(networkPath, providerHost, method)
-
-	// Ensure the replacer has the correct payload for consistent logging
-	repl.Set(RequestBodyKey, payload)
-
-	// Pass the address of respStatus to processBlockNumberResponse
-	// processBlockNumberResponse checks for respStatus >= 400
-	blockNumber, _, processingError := networkObj.processBlockNumberResponse(respBody, &respStatus)
-	if processingError != nil {
-		// Log the failed async processing with detailed information
-		logFailedAttempt(&LogFailedAttemptParams{
-			Reason:              "Async health check processing failed",
-			Logger:              d.logger,
-			NetworkPath:         networkPath,
-			FailedAttemptNumber: 1, // Single attempt for async processing
-			MaxAttempts:         1, // Total attempts is always 1 for async processing
-			StatusCodeOfFailure: respStatus,
-			Error:               processingError,
-			Replacer:            repl,
-			ParsedReqBody:       jsonRPCReq,
-			RawResponseBody:     respBody,
-		})
-		return
-	}
-
-	block, err := networkObj.getBlockByNumber(blockNumber)
-	if err != nil {
-		// Create a new context specifically for the getBlockByNumber call that failed
-		// This ensures the logging shows the correct method and parameters for the failed call
-		getBlockMethod := networkObj.GetBlockByNumberMethod
-		if getBlockMethod == "" {
-			getBlockMethod = DefaultGetBlockByNumberMethod // Default fallback
-		}
-
-		// Create context for the getBlockByNumber call
-		getBlockRepl, getBlockJSONRPCReq, _ := createGetBlockByNumberRequestContext(networkPath, providerHost, getBlockMethod, blockNumber, networkObj)
-
-		// For getBlockByNumber errors, we don't have a JSON-RPC error from the original response
-		// since this is a separate internal call
-		logFailedAttempt(&LogFailedAttemptParams{
-			Reason:              "Get block by number failed",
-			Logger:              d.logger,
-			NetworkPath:         networkPath,
-			FailedAttemptNumber: 1,   // Single attempt for async processing
-			MaxAttempts:         1,   // Total attempts is always 1 for async processing
-			StatusCodeOfFailure: 200, // getBlockByNumber is an internal call, assume 200 for the original response
-			Error:               err,
-			Replacer:            getBlockRepl,
-			ParsedReqBody:       getBlockJSONRPCReq,
-			RawResponseBody:     nil, // No response body for internal getBlockByNumber calls
-		})
-		return
-	}
-
-	// save the block number to the network object's history
-	networkObj.AddNetworkBlockEntry(blockNumber, block) // Add the block number and block hash to the network object's history as long as its the the latest block number
-	d.logger.Debug("Goroutine: HCMethod matched, successfully processed block number and added to network history", zap.Int64("block_number", blockNumber), zap.String("network", networkPath))
 }
 
 // createHealthCheckRequestContext creates synthetic request context for health checks
